Start thinking about a sqlparse fuzzer.
This is another alternative to regexps and sqlalchemy... At the
moment this code is just a plaything to see if it works at all,
but at the moment it can read all of user 001, detect the create
statements, and then dump it out with a zero sized diff. This
gives me hope.